refactor: 점수 계산 역할 위치 수정
